Adds a configuration for embed mode, enabled by default, to allows viz filters or not.
Sometimes we may want to simply embed a visualization within a page using an iframe as a "read only" chart and where there is no need for the user to filter it.

And sometimes it simply does not make sense to let the user filter it, for exemple if you embed a pie chart out of any context, it is not necessary to give the end user the possibility to filter on a category since the change of context does not impact anything else.
(cherry picked from commit 7abb5a5)